requirements.txt
nltk
searchtweets-v2==1.0.6